Rename to common.h
