+++ src/kernels/common.h
